requirements (2).txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
passlib
python-jose
pyotp
email-validator
